requirements.txt
# requirements.txt
boto3>=1.26.0
opensearch-py>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0